Project setup
Created a .gitignore file in the root folder and a .env file in the
client folder. Installed eslint as a dev dependency and created a
.eslintrc.json file with three custom rules. Installed react-query,
react-router-dom, and @chakra-ui/react. Deleted unnecessary files from
create-react-app, which are App.css, App.test.js, and logo.svg, and
removed the imports of those files.

Imported Provider components into index.js and wrapped them around
App component.