Changes to .eslint.json, started making components
Formatting rules were added to .eslint.json and the formatting for all files in this project were updated accordingly

A shared folder was created. This folder will contain all of the components of which multiple instances will exist. A Button.js functional component was created. This compoonent will be the common button component.

A pages folder was created. This folder will contain all of the components that will act as a page for the web app. That is, each component in this folder will be rendered once and only once, and whose direct parent will be App.js. Home.js was created and will be the home page for the app.

react-router-dom components were imported into App.js to wrap around the Home.js component.

A custom theme for Chakra-UI was created in index.js. This overrides the default breakpoints that Chakra-UI uses for media queries.