Created inputs
I created the general input component. The type, as well as other attributes, are specified by a prop.

I also edited the Home.js file to move the buttons higher toward the navbar when on screens smaller than 'medium' (this will apply to mobile devices).

I also slightly changed the style of the buttons to make them smaller on a mobile device and reduced the font weight. These styles were previously passed into the button as props, but I moved them into the ButtonTheme.js file.