Created login page
A login page was created. It uses the same components and layout that were used in the signup page. The login page component was then wrapped around a Route component in App.js. I also changed the fonts.js file so that all anchor (a) elements use the Rubik font family. However, this is not working and so I created another branch named font-fix to attempt to fix this.