Changed width of VNavbar and w of Input.js
I widened the width of the vertical Navbar, VNavbar, to be 20rem so that a search button will fit beside both inputs. I also changed the w prop in Input.js to be equal to props.w, rather than w='100%', so that w can be changed from any other component (by passing in the desired value as a prop).