Created vertical navbar
I created a vertical navbar, as well as a shared component called InputGroupRight. This is for inputs with an icon button inside of the input, but on the right-hand side. I installed a package called react-icons to use for the two search bars on the vertical navbar.

This vertical navbar will be used for screen sizes greater than or equal to 990px (width), but otherwise, I will use a horizontal navbar that is identical to the HNavbar already created, except with a hamburger menu option or something similar.

I also removed the default background color, buttonColor, in ButtonTheme.js, as IconButton is composed with a Button, and all of my IconButtons were appearing with a buttonColor background and I was unable to override this. To ensure that the buttons had that same buttonColor background, I added it as a prop on the pages where the Button is rendered.