Created large-screen layout for LoggedIn.js
The layout for a large screen will be a vertical navbar on the left side, spanning the entire height of the screen. There will be navigation buttons in the top right corner of the screen,
and the recipe content will be below those navigation icons, and to the right of the navbar.

This layout was achieved using a grid. I coloured the grid items so that I can see where each individual section is. However, I have already rendered the vertical navbar. Note that for a mobile screen the navbar will be at the top and will be horizontal. It will contain the navigation buttons in some type of menu, and the content will be below that. That layout will be created later.