Editted VNavbar
I redid the layout in the VNavbar so that the items are bunched more closely together and nearer to the top. I also removed an unncessary VStack component that wrapped around the Add Recipe link and the inputs for searching people and recipes. I use the spacing prop to handle their spacing now. I also removed the link for Profile Settings, since I will be using a navigation button for that.