Refactoring of LoggedIn.js and other components
LoggedIn was refactored to include layouts for large screens, medium screens, and small screens. HNavbar was also refactored to include navigation buttons in it for medium screens.
The Add New Recipe link was removed from VNavbar and instead, I added a navigation button for adding a new recipe. I also added in a custom breakpoint called smd, which stands for small-medium. It is 39rem, which is halfway between 30rem (sm) and 48rem (md).