New styles
I added tabs in RecipeDrawer - one tab for description, one tab for the ingredients, etc. I also moved the add and delete buttons next to the headers, and I made it so that when you have a number of ingredients or reciep steps such that there is a vertical scroll bar, the items scroll behind the headers. So the headers act similar to a fixed navbar.