Removed layout for medium screens
I removed the layout for screens between 39rem and 62rem. This layout was only going to cause future problems. I instead use the layout for mobile in place of it. It looks nicer and is more user-friendly. Because of this, there is no longer really a horizontal nav bar. It is really just a bar with the logo on it. So I renamed HNavbar.js to HLogoBar.js and updated all imports and uses of the component accordingly.