Signup form displays errors now
If a network error occurs (for example, if the user tries to create an account but the server is down), then some text describing the error will appear below the form.

The form also now shows it is submitting by using the typical loading animation on the submit button.

I also removed the error catching on api.js for submitting the form. This was done because catching the error would result in a resolved promise being returned, which would cause react-query to think that the POST request was successful. Not catching the error signals to react-query that an error occured. The fact that an error occured is reflected in the boolean value isError, which is returned by useMutation. I then conditionally render the error text, depending on the value of isError.