Passport authentication was created
I used passport and passport-local to allow the user to log in. So far only the server can accept the login credentials. That is, the user cannot currently log in using the web app. However, they can log in using the curl api or Postman, for example. A session using cookies is also created for the user. I used express-session and connect-mongo for this.