Having login form submit
The login form can now submit and log a user in. I also changed the route to access the logged in page to be /users/:id. I still have to program a check to the server, upon trying to access /users/:id, that the user is indeed logged in. I will do that next.