Accessing protected routes
I created a middleware that will be ran whenever the user tries to access a protected route. If req.user exists, meaning the user is logged in and there is a valid session, then we continue onto the next middleware function. Otherwise, an error message is returned and the request-response cycle is ended.