Save recipe route has been created
I edited the user model so that it now has a data field called recipes, which by default is an empty array (of mongoose Maps - which is just a javascript object). I also wrote the controller function that reads the recipe information from the request body and saves it as a recipe in the aforementioned array.