Saving recipe api function created
I created the api function to submit a POST request to an api endpoint. I also created that api endpoint to receive the request, verify that the user is logged in, and if so, save that recipe in the user's profile in MongoDB. If the user is not logged in, an error is generated and the user is told that they have to log in.

I also write an api function called getUserID that will simply fetch the user's MongoDB id - provided they're logged in. This is necessary to create the URL for posting the recipe. More specifically, when the user tries to save a recipe, a POST request is made to /users/:id/recipes. However, the :id parameter has to actually be replaced with the user's ID in order for the post request to be successful. So the ID is first retrieved with the function getUserID.