Code cleaned up
This branch was created with the intention of cleaning up and reorganizing some of my code. I deleted imports that weren't being used, and I also deleted redundant imports. For example, in each React component I import React from 'react'. In some of those React components, I also imported certain React hooks, such as useState, useEffect, etc. There is no need to import these because they are included in the React module that I imported already. So, for example, if there was a useState in my code, it was changed to React.useState.

I also moved all functions in customMiddleware.js and moved them into authController.js. It makes more sense for those files to be there. Every controller function that I write is technically custom middleware, so having a separate file called customMiddleWare didn't make sense.

I also prepended all routes with api/. So for example, the route http://example.com/users is now http://example.com/api/users. I did this to make it more clear that this is purely an API. I also updated the URLs in the fetch requests in my api.js file (in the client folder) accordingly.