Redirects to profile page if logged in
I made it so that if the user is logged in, they are redirected to their profile page if they try to access the home page, the sign up page, or the login page. I did this by making the login api call save the user id to localStorage. I then wrote a customHook called useAuthentication which just checks if a userId exists in localStorage. It returns true if so, and false otherwise. The home page, sign up page, or login page, then render their usual components if and only if the user is NOT logged in. Otherwise it redirects the user to their profile page.

At first glance this may seem insecure, because if somebody were to obtain a user's user id and enter it into their localStorage, the app would then direct them to that user's profile page. However, the resources required to properly load a user's profile page such as their profile picture, their recipes, etc, are all loaded via api calls that require authentication via a session cookie delivered through the browser. Simply writing the user's user id into the local storage will not reproduce the session cookie. So the app is still secure.

I also changed the isloggedin function in authController.js to use the passport function req.isAuthenticated, and I fixed a small mistake in recipeController.saveRecipe: The username field for the updatedUser that was being saved in the MongoDB database was written as "name". I corrected it by changing it to "username".