Logging out user
The ability for the user to logout had been implemented. The api endpoint was created to log the user out. Logging the user out entails first deleting the user object from req.user. With that done, it is impossible for protected routes to be accessed. Secondly, the session is deleted from the MongoDB session store. And lastly, the cookie is deleted in the browser.

I also, of course, wrote the function that makes the api call to the server. Once the user is logged out (logged out being defined as the api call was successful), the localStorage is cleared of the 'userId' key. The user is then redirected back to the home page.