Logout error handler
I added an error handler for the logging out function. If the api call made to log the user out is unsuccessful for any reason, a toast appears on the screen informing the user that there was a network error and that they must clear all of their cookies in order to log out.