Made Redux store the image along with image URL
I made Redux store the image file along with the image URL. The reason for this is because when the user eventually submits the recipe to the server so that it can be saved, the server requires the actual image file to send to the AWS S3 bucket. So at first I tried to just save the image file INSTEAD OF the object url. However, this slowed down the rendering of the photo when the user chose the photo for their recipe. The solution is to store both the file and the object URL. The object URL will be used to display the photo to the user when they're selecting for their recipe. This is fast and is a good user experience. However, when the user submits the recipe to be saved, the image file itself will be used.

I made the changes in photoSlice.js to reflect this change of strategy. I also subtracted 1 from each array index in the photoSlice.js reducers, as photoNum starts counting from 1, but of course in arrays we count from 0.

Note that the image file is a nonserializable value, and hence Redux gives a warning in the browser console to avoid this. Redux only wants you to save serializable data in its store because it only detect changes in serializable data, and hence can only rerender properly if serializable data is stored. However, this is not an issue for the strategy that I am using. I store the image, which is nonserializable, but I also store the object URL, which IS serializable. Since the photos displayed to the user in the RecipeDrawer use the object URL, then Redux will rerender just fine if the user adds, updates, or deletes a photo. It is important to note that the image files, the nonserializable data, ARE updating in the Redux store when the user adds, updates, or deletes a photo. It is just that Redux won't rerender anything that depends on the image files - which is not an issue for us.