Prevent access to LoggedIn if not logged in
I used the useAuthentication hook that I wrote to redirect the user to the log in page if they are not authenticated but try to access /users/:id. This is just for user experience reasons. There is no security risk if I don't do this, because even if LoggedIn.js is loaded without the user being authenticated, the resources that display protected information won't be displayed unless the user is actually authenticated via a cookie.