Using the FormData API
I am now slowly creating the feature to add photos to your recipe (and ultimately store them in the S3 bucket). However, the Content-Type header on the fetch request has to be multipart/form-data, rather than application/json. I also have to use the FormData api to add the photos to the request body. Note that because I am using the FormData api, I omit Content-Type in the fetch request, as the browser will automatically detect it.

My Express backend was not set up to parse a request whose Content-Type is multipart/form-data. This is one fo the uses of Multer. I use Multer as a middleware, which allows my other middleware functions to also work (such as sanitizing the form input).