RecipeDrawer can upload photos to S3 Bucket
I now use the FormData api to create a form, which the client then submits to the server. I did this so that I can attach photos to the request object. I use Multer on the server to parse the multipart form and upload the photos to the S3 Bucket.

Note that I upload the photos to the bucket first, because I will save the photo keys in MongoDB. The keys will be used to retrieve the photos from the S3 bucket whenever the user wants to load a recipe. So I HAVE to upload the photos to the S3 bucket first so that I can get the photo keys.

What happens if I successfully upload the photos but there is an error when saving the text portion of the recipe to MongoDB? I will write a middleware function to remove the photos from the S3 bucket.