Added validation for recipe name
I validate the recipe name in Multer's fileFilter function (whcih is ran before the photos are uploaded). The upload.array function that is used to upload the photos to the S3 bucket will not execute if it is placed after the express-validator functions. This is because the form submitted is multipart and Express cannot parse it. Multer must be ran first so that it can populate req.body. The express-validator functions will be ran after that.

Note that express-validator still sanitizes any input (including the recipe name) after Multer populates req.body.

I also had the server (as opposed to the client) add a unique ID to each recipe. The reason for this is because if someone is interacting with the back end with an api client, then a recipe ID won't be added to each recipe.