Delete photos if there is an error
Multer will upload the photos as long as the recipe name field passes validation, i.e. it exists. However, what happens afterward if for some reason express-validator fails to validate req.body? Or what happens if there is an error when uploading the recipe to MongoDB? That means that the recipe won't be saved to MongoDB, but the photos were still uploaded to the S3 bucket. This is a waste of space. I have added the necessary code to remove the photos from the S3 bucket should either of those two scenarios occur.