Fixed issue with Save button disappearing
When using the app on my cellphone (a Samsung J7 and using the native Samsung internet browser), the Save button in the RecipeDrawer would become invisible when you clicked on it to save a recipe. However, the button worked fine when saving a recipe on my laptop. This seems to be a bug in Chakra-UI. I will have to submit an issue to their Github.

In the meanwhile, to get around it, I avoided setting the variant prop on the button to 'drawerButton', which is a custom theme that I created. Instead, I created an object in the RecipeDrawer component called drawerButtonProps, with the exact same styles that defined 'drawerButton'. Then I used to spread operator to apply these props to the Save button.

That fixed the issue.