Showing potential errors to user
I made it so that a toast pops up showing the error message, if any, the user receives back from the server when submitting their recipe. For the sake of consistency, all error messages that are sent back are sent in JSON format with a key called 'errorMessage'. This makes it easy to check if the recipe was saved (check for the existence of the key errorMessage in the response), and it makes the error message easily accessible (it is the value of errorMessage).

I also changed the formatting of some of the errorMessages that are sent back. I made these changes in recipeController.js and authController.js. All I did was replace any template literals strings (strings with backticks surrounding them) with strings that are encompassed by single quotes for all errorMessages. I did this because the formatting of the string being returned had \n and \t escape characters in them.

I also added in some try-catch blocks for the api call and for the response received from the api call. I did this to catch any errors that the api call may encounter.